Implemented [uid] endpoint for observation units
